creating sorts.go
Moving some code from msgs.go to sorts.go
