+++ eod/msgs.go
@@ -17,14 +17,6 @@ var seps = []string{
 	"plus",
 }
 
-func makeListResp(start, join, end string, vals []string) string {
-	if len(vals) == 2 {
-		return fmt.Sprintf("%s %s %s %s%s %s", start, vals[0], join, vals[1], end, types.RedCircle)
-	} else if len(vals) > 2 {
-		return fmt.Sprintf("%s %s, %s %s%s %s", start, strings.Join(vals[:len(vals)-1], ", "), join, vals[len(vals)-1], end, types.RedCircle)
-	}
-	return ""
-}
 func (b *Bot) PingCmd(c sevcord.Ctx, opts []any) {
 	c.Acknowledge()
 	ping := b.s.Dg().HeartbeatLatency().Microseconds()
@@ -34,45 +26,6 @@ func (b *Bot) PingCmd(c sevcord.Ctx, opts []any) {
 		c.Respond(sevcord.NewMessage(c.Author().Mention() + " 🏓 Pong! Latency:**" + fmt.Sprintf("%f", seconds) + "s**"))
 	} else {
 		c.Respond(sevcord.NewMessage(c.Author().Mention() + " 🏓 Pong! Latency:**" + fmt.Sprintf("%f", milliseconds) + "ms**"))
-	}
-}
-func (b *Bot) getElementId(c sevcord.Ctx, val string, showerr bool) (int64, bool) {
-	var id int64
-	var err error
-	id, err = strconv.ParseInt(strings.TrimPrefix(strings.TrimSpace(val), "#"), 10, 64)
-	if err == nil && strings.HasPrefix(val, "#") {
-		err = b.db.QueryRow("SELECT id FROM elements WHERE id=$1 AND guild=$2", strings.ToLower(strings.TrimLeft(strings.TrimSpace(val), "#")), c.Guild()).Scan(&id)
-	} else {
-		err = b.db.QueryRow("SELECT id FROM elements WHERE LOWER(name)=$1 AND guild=$2", strings.ToLower(strings.TrimSpace(val)), c.Guild()).Scan(&id)
-	}
-	if err != nil && showerr {
-		b.base.Error(c, err, "Element **"+val+"** doesn't exist!")
-		return 0, false
-	} else if err != nil && !showerr {
-		return 0, false
-	}
-
-	return id, true
-}
-func getSort(input string) string {
-
-	switch strings.TrimSpace(strings.ToLower(input)) {
-	case "creator":
-		return "creator"
-	case "name":
-		return "name"
-	case "id":
-		return "id"
-	case "created on", "createdon":
-		return "createdon"
-	case "treesize", "tree size":
-		return "treesize"
-	case "length":
-		return "length"
-	case "found":
-		return "found"
-	default:
-		return "id"
 	}
 }
 
@@ -162,35 +115,7 @@ func (b *Bot) textCommandHandler(c sevcord.Ctx, name string, content string) {
 			return
 		}
 		parts := strings.Split(content, " ")
-		Lbsort := ""
-		switch strings.ToLower(parts[0]) {
-		case "made":
-			Lbsort = "made"
-		case "imaged", "img":
-			Lbsort = "img"
-		case "votes":
-			Lbsort = "voted"
-		case "signed":
-			Lbsort = "signed"
-		case "colored", "colour", "coloured", "color":
-			Lbsort = "color"
-		case "catsigned":
-			Lbsort = "catsigned"
-		case "catimg", "catimage":
-			Lbsort = "catimg"
-		case "catcolor":
-			Lbsort = "catcolor"
-		case "querysigned":
-			Lbsort = "querysigned"
-		case "queryimage", "queryimg":
-			Lbsort = "queryimg"
-		case "querycolor":
-			Lbsort = "querycolor"
-		case "found", "":
-			Lbsort = "found"
-		default:
-			Lbsort = "found"
-		}
+		Lbsort := getLbSort(strings.ToLower(parts[0]))
 		b.pages.Lb(c, []any{Lbsort, nil, nil})
 
 	case "p", "products":
